+++ app/src/main/kotlin/ktx/app/application.kt
@@ -5,52 +5,6 @@ import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.InputProcessor
 
 /**
- * Abstract implementation of [ApplicationListener] similar in scope to [com.badlogic.gdx.ApplicationAdapter]. Provides
- * empty implementations of most optional methods. Handles screen clearing (with black color) and rendering on a fixed
- * step value.
- * @param fixedTimeStep minimum time (in seconds) between two consequent render calls. Might not match the delta time
- *    reported by [Gdx.graphics]. Defaults to 1/60.
- * @param maxDeltaTime maximum time (in seconds) stored by the application listener for fixed time step calculations.
- *    Defaults to 1 second.
- */
-abstract class KotlinApplication(protected val fixedTimeStep: Float = 1f / 60f,
-                                 protected val maxDeltaTime: Float = 1f) : ApplicationListener {
-  /** Internal control variable used to sure fixed time step durations. In seconds. Might not match the actual time
-   * since the last [render] call in case of subsequent [render] calls on devices unable to run the application at the
-   * chosen time step rate.*/
-  protected var timeSinceLastRender = 0f
-    private set
-
-  override fun resize(width: Int, height: Int) {
-  }
-
-  override final fun render() {
-    timeSinceLastRender = Math.min(timeSinceLastRender + Gdx.graphics.rawDeltaTime, maxDeltaTime)
-    while (timeSinceLastRender >= fixedTimeStep) {
-      timeSinceLastRender -= fixedTimeStep
-      clearScreen(0f, 0f, 0f, 1f)
-      render(fixedTimeStep)
-    }
-  }
-
-  /**
-   * Called by [render] function when the raw delta time reported by current [Gdx.graphics] reaches the chosen
-   * [fixedTimeStep] value. Ensures smooth rendering on fixed time step value.
-   * @param delta estimate of time since last method call.
-   */
-  abstract fun render(delta: Float)
-
-  override fun pause() {
-  }
-
-  override fun resume() {
-  }
-
-  override fun dispose() {
-  }
-}
-
-/**
  * Wrapping interface around [com.badlogic.gdx.ApplicationListener]. Provides no-op implementations of all methods,
  * making them optional to implement.
  */